Task 101 v7

+++ 0x00-hello_world/101-quote.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	int str = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
+	short str = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
 
 	putchar(str);
 	/* return an int */
